Final push to dev before publication
>Fixed friendly fire, now being very rare.
>Fixed bug in code which prevented Villagers from properly joining a
team, with confusion ensuing therefrom.
>Fixed GUI to be pretty and functional, both Baton and Villager
>Changed/Fixed how villagers treat reputation:
>>Each villager contributes to player rep of whole village based on
average
>>Each villager will change his reputation of player in accordance with
the change in the whole village, if the reputation is higher than the
villager's current reputation for the player, or the change is negative.
>>After the village reputation for player has been calculated, and the
individual villagers have each had their personal reputation for the
player updated, a new average is calculated, and that is used to update
the "reference reputation" for each villager.
>Made changes to Ranged shooting AI:
>>Added optional Head-facing debug info for when Entity shoots after
sudden head movement (long time issue)
>>Removed target-seen time requirement (time seen irrelevant to loading
a weapon)
>>If Villager shoots a bow, creates arrow based on bow as before, but
now custom to each arrow used.
>Added movement speed for villagers as a config var
>Added more info to VilPlayerDeal debug.